exp09_common_subexpressions is updated
